+++ src/internal/arguments.rs
@@ -1,6 +1,6 @@
-use std::fmt::Debug;
+use std::fmt::{Debug, Display};
 
-pub trait Arguments: Debug {}
+pub trait Arguments: Display + Debug {}
 
 #[cfg(test)]
 pub(crate) use self::mock::*;
@@ -8,9 +8,16 @@ pub(crate) use self::mock::*;
 #[cfg(test)]
 mod mock {
     use super::Arguments;
+    use std::fmt;
 
     #[derive(Debug)]
     pub(crate) struct ArgumentsMock;
 
     impl Arguments for ArgumentsMock {}
+
+    impl std::fmt::Display for ArgumentsMock {
+        fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+            write!(f, "mock")
+        }
+    }
 }