+++ src/internal/fmt/debug.rs
@@ -1,5 +1,6 @@
 use std::fmt::{self, Debug};
 
+/// A wrapper around a [`MaybeDebug`] type that implements [`Debug`].
 pub struct MaybeDebugWrapper<'a>(pub &'a dyn MaybeDebug);
 
 impl<'a> Debug for MaybeDebugWrapper<'a> {
@@ -8,23 +9,9 @@ impl<'a> Debug for MaybeDebugWrapper<'a> {
     }
 }
 
-pub struct MaybeDebugExtWrapper<'a>(pub &'a dyn MaybeDebugExt);
-
-impl<'a> Debug for MaybeDebugExtWrapper<'a> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        MaybeDebugExt::fmt(self.0, f)
-    }
-}
-
 /// This trait makes every type [`Debug`] by falling
 /// back to "?" when [`Debug`] is not implemented.
 pub trait MaybeDebug {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result;
-}
-
-/// A wrapper around [`MaybeDebug`] for container types such as [`Box`],
-/// because specialization does not allow impls for `Box<T>`.
-pub trait MaybeDebugExt {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result;
 }
 
@@ -43,14 +30,5 @@ where
 {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "{:?}", self)
-    }
-}
-
-impl<T> MaybeDebugExt for Box<T>
-where
-    T: MaybeDebug + ?Sized,
-{
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        MaybeDebug::fmt(&**self, f)
     }
 }