+++ src/internal/matcher/any.rs
@@ -1,15 +1,16 @@
 use super::ArgumentMatcher;
 use std::fmt::Write;
-use std::fmt::{self, Debug};
+use std::fmt::{self, Display};
 
 /// Creates a new `ArgumentMatcher` that matches any value.
 pub fn any() -> AnyArgumentMatcher {
     AnyArgumentMatcher
 }
 
+#[derive(Debug)]
 pub struct AnyArgumentMatcher;
 
-impl Debug for AnyArgumentMatcher {
+impl Display for AnyArgumentMatcher {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.write_char('*')
     }