+++ changelog.md
@@ -12,3 +12,8 @@
 
 ## 0.4.0
 - Ordered expectations are now available using `mock.expect_<method_name>_calls_in_order()`
+
+## 0.4.1
+- The `Debug` implementation of an argument's type is now correctly used when printing the expected calls for a method.
+- Internal types now have separate `Debug` and `Display` implementations.
+- Some unused exports have been removed from `mockiato::internal`. This is treated as a non-breaking change as no consumer should directly depend on these types.